Add: src/apis/signup.jsをTypeScript対応させた
