+++ frontend/src/apis/ranking.ts
@@ -5,14 +5,10 @@ import { rankings } from '../urls/index';
 // postの第3引数にwithCredentials: trueを指定することで、
 // API(Rails)と通信する際にデータにcookieを含めることができる
 export const getRanking = async (): Promise<any> => {
-  try {
-    const response = await axios.get(
-      rankings,
-      { withCredentials: true }
-    );
-    axios.defaults.headers.common['X-CSRF-Token'] = response.headers['x-csrf-token'];
-    return response.data;
-  } catch(e) {
-    throw e;
-  }
+  const response = await axios.get(
+    rankings,
+    { withCredentials: true }
+  );
+  axios.defaults.headers.common['X-CSRF-Token'] = response.headers['x-csrf-token'];
+  return response.data;
 };
